backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend para o Otimizador de Currículos com IA"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.0",
    "pypdf>=4.2.0",
    "spacy[pt-core-news-sm]>=3.7.0",
    "google-generativeai>=0.3.0",
    "aiofiles>=23.0.0",
    "agno>=1.7.11",
    "pytest>=8.4.1",
    "python-magic>=0.4.27",
    "email-validator>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
